backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
telethon==1.34.0
apscheduler==3.10.4
sqlalchemy[asyncio]==2.0.31
aiosqlite==0.20.0
httpx==0.27.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
pysocks==1.7.1
